requirements.txt
imageio==2.31.5
numpy==1.24.4
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
pandas==2.1.3
Pillow==10.0.1
Pillow==10.1.0
pytorchvideo==0.1.5
scikit_learn==1.0.2
torch==2.1.1
torchinfo==1.8.0
torchvision==0.16.1
